move wrapper tasks to tasks/wrapper.py
